% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_CPTAC_harmonized_data.R
\name{access_harmonized_CPTAC_data}
\alias{access_harmonized_CPTAC_data}
\title{access_harmonized_CPTAC_data}
\usage{
access_harmonized_CPTAC_data(tumors, data_types)
}
\arguments{
\item{tumors}{vector of strings with tumor names. The user can choose from these names: "Brca", "Ccrcc", "Coad", "Gbm", "Hnscc", "Lscc",
"Luad", "Ov", "Pdac", "Ucec".}

\item{data_types}{vector of strings containing the data types to extract. The options could be "phospho", "prot", "transc" and "mut"
(respectively phosphoproteomics, proteomics, transcriptomics and somatic mutations)}
}
\description{
This function provides access to harmonized CPTAC data for 10 tumor types and 3 omics per tumor type. The data has been harmonized using the
PatientProfiler function \emph{omics_update}. Users can access the data to use them in the other functions for protein activity inference and
for network creation. It creates local variables with the dataframe(s) selected by the user.
}
\examples{
# Extract Phosphoproteomic and Proteomic dataframes for Brca and Ccrcc
access_harmonized_CPTAC_data(tumors = c("Brca", "Ccrcc"), data_types = c("phospho", "prot", "mut"))

# Six local variables, Brca_phospho, Brca_prot, Brca_mut, Ccrcc_phospho, Ccrcc_prot and Ccrcc_mut will be generated.

}
