% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_multiplicity.R
\name{add_multiplicity}
\alias{add_multiplicity}
\title{add_multiplicity}
\usage{
add_multiplicity(df, peptide_col)
}
\arguments{
\item{peptide_col}{name of the column containing the peptide sequence}

\item{phospho_df}{tibble, the phosphoproteomics dataframe}
}
\value{
phosphoproteomics dataframe with an added 'Multiplicity' column that counts phosphosites in the peptide sequences
}
\description{
add_multiplicity
}
\examples{
sample_df <- data.frame(
  Gene_name = c("ACIN1", "ZZEF1", "AAK1")
  Peptide = c("sRsRsTPVRR", "LLsFRsMEEAR", "SATTTPSG*PR"),
  Value = c(0.1, 0.2, 0.3)
)

modified_df <- add_multiplicity(sample_df, peptide_col = "Peptide")
}
