% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{aes_dataset}
\alias{aes_dataset}
\title{Split dataset into geoms}
\usage{
aes_dataset(dt, aes_map)
}
\arguments{
\item{dt}{full dataset of all patients' data}

\item{aes_map}{see \code{\link{aes_map}}}
}
\description{
Split dataset into geoms
}

