% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_signature_inputs.R
\name{check_signature_inputs}
\alias{check_signature_inputs}
\title{Check input format for test_signature_on_dataset}
\usage{
check_signature_inputs(
  gene_signatures,
  transcriptomic_data,
  clinical_data = NULL,
  enrichment_result = NULL
)
}
\arguments{
\item{gene_signatures}{List of named vectors}

\item{transcriptomic_data}{Dataframe with gene_name column}

\item{clinical_data}{Dataframe with patient_id, serial_time, status columns (can be NULL)}
}
\description{
Check input format for test_signature_on_dataset
}
