% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_multiple_peptide.R
\name{clean_multiple_peptide}
\alias{clean_multiple_peptide}
\title{clean_multiple_peptide}
\usage{
clean_multiple_peptide(phospho_df, peptide_col)
}
\arguments{
\item{phospho_df}{tibble, the phosphoproteomics dataframe}

\item{peptide_col}{name of the column containing the peptide sequence}
}
\value{
phosphoproteomics dataframe with one Peptide, the longest, for each phosphosite
}
\description{
clean_multiple_peptide
}
\examples{
sample_phospho_df <- data.frame(
  Gene_nname = c("UNKL", "SRGAP3", "NSUN5"),
  Peptide = c("AAAAALSGsPPQTEK;AAAAALSGsPPQTEKPTHYR;", "AAACPSsPHK", "AAAGACtPPCT"),
  Value = c(0.1, 0.2, 0.3)
)

cleaned_df <- clean_multiple_peptide(sample_df, peptide_col = "Peptide")
}
