% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort_networks.R
\name{create_cohort_networks}
\alias{create_cohort_networks}
\title{create_cohort_networks}
\usage{
create_cohort_networks(
  trans_dir = NULL,
  prot_dir = NULL,
  phospho_dir = NULL,
  act_dir = NULL,
  desired_phenotypes = c(),
  pheno_distances_table = NULL,
  mut_file = NULL,
  output_dir = "./Networks_output/",
  save_all_files = FALSE,
  PKN_options = list(),
  naive_options = list(),
  carnival_options = list(),
  phenoscore_options = list(),
  format_options = list()
)
}
\arguments{
\item{prot_dir}{string, path to cohort proteomics data folder, default \code{NULL}.}

\item{phospho_dir}{string, path to cohort proteomics data folder, default \code{NULL}.}

\item{act_dir}{string, path to cohort proteomics data folder, default \code{NULL}.}

\item{desired_phenotypes}{SIGNOR phenotypes vector to infer the activity and include in the model; default:  \code{NULL}.}

\item{pheno_distances_table}{character, any character except NULL to internally compute it; default:  \code{NULL}.}

\item{mut_file}{string, path to cohort mutations .tsv file, default \code{NULL}.}

\item{output_dir}{string, path to network folder; default \code{'./Networks_output/'}.}

\item{save_all_files}{Boolean, if TRUE it will save 13 files per patient, FALSE, just 4; default \code{FALSE}.}

\item{PKN_options}{check \emph{initialize_net_default_params} documentation, in the \verb{$PKN_options} part.}

\item{naive_options}{check \emph{initialize_net_default_params} documentation, in the \verb{$naive_options} part.}

\item{carnival_options}{check \emph{initialize_net_default_params} documentation, in the \verb{$carnival_options} part.}

\item{phenoscore_options}{check \emph{initialize_net_default_params} documentation, in the \verb{$phenoscore_options} part.}

\item{format_options}{check \emph{initialize_net_default_params} documentation, in the \verb{$format_options} part.}
}
\value{
No return value. The final results are saved in \emph{RDS} and \emph{sif} files in the output directory.
}
\description{
This function loops on the folder of omics and activity data of a cohort of patients
creating a network of each of the patients from mutations to desired phenotypes.
For the files generated for each patient, refer to \emph{create_network} function documentation.
}
\examples{

PKN_options = list()
naive_options = list()
carnival_options = list()
phenoscore_options = list()
format_options = list()

create_cohort_networks(trans_dir = './Trans_test/',
                      prot_dir = './Prot_test/',
                      phospho_dir = './Phos_test/',
                      act_dir = './Act_test/',
                      mut_file = 'mutations.tsv',
                      output_dir = './Networks_output/',
                      desired_phenotypes = c('APOPTOSIS', 'PROLIFERATION'),
                      pheno_distances_table = TRUE,
                      PKN_options = list(direct = FALSE),
                      naive_options = list(layers = 2, max_length = c(1,4)),
                      carnival_options = list(solver = 'cplex', carnival_type = 'inverse'),
                      format_options = list(optimize_on_phenotypes = TRUE, vis_cytoscape = FALSE))
}
