% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_naive_network.R
\name{create_naive_network}
\alias{create_naive_network}
\title{create_naive_network}
\usage{
create_naive_network(
  PKN,
  sources,
  activities,
  output_dir,
  naive_network_parameters = list()
)
}
\arguments{
\item{PKN}{dataframe obtained from the function get_PKN representing PKN interactions}

\item{sources}{dataframe of perturbed nodes (e.g., patient mutations), starting point of the model}

\item{activities}{dataframe of proteins inferred activity to connect to the starting point}

\item{output_dir}{string, output naive network directory}

\item{naive_network_parameters}{A list of parameters for naive network construction. Available parameters are:
\itemize{
\item \code{layers}: integer, specifiying the number of layers > 0 and < 4; default  \code{2}
\item \code{max_length} a vector of int or an int, length == layers; default  \code{c(1,4)}
\item \code{connect_all} Boolean, whether connecting intermediate nodes of shortest path; default  \code{TRUE}
\item \code{naive_path} path of the folder where naive network files are saved; default \code{./Network_output/Naive_}
}}
}
\value{
igraph object representing the naive network
}
\description{
Generates hierarchical (multi)layered network from source nodes (perturbed nodes) defining
layers by distinguishing inferred proteins by molecular function.
}
\details{
\emph{Results output}
The function:
\itemize{
\item returns an \emph{igraph object} representing the naive network
\item creates internally an \emph{RDS file} and a \emph{SIF file} of the naive network at \code{output_dir} directory
}
}
\examples{
omics_list <- list(transcriptomics, proteomics)
PKN = get_PKN(omics_list = omics_list)
create_naive_network(PKN, sources, activities, naive_network_parameters = list(layers = 2, max_length = c(1,4)))
}
