% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_patient_id}
\alias{extract_patient_id}
\title{extract_patient_id}
\usage{
extract_patient_id(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
it returns a string representing patient id
}
\description{
Extract patients_IDs from PatientProfiler directories structures
}
\examples{
extract_patient_id('Prot_Patient_X11.xlsx')

}
