% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_protein_activity.R
\name{extract_protein_activity}
\alias{extract_protein_activity}
\title{extract_protein_activity}
\usage{
extract_protein_activity(
  prot_file = NULL,
  trans_file = NULL,
  phospho_file = NULL,
  tf_params = list(),
  kin_params = list(),
  phosphoscore_params = list(),
  phosphoscore_noseqwin_params = list(),
  output_dir = "Activities"
)
}
\arguments{
\item{prot_file}{a string (optional) indicating the proteomics data file (default NULL)}

\item{trans_file}{a string (optional) indicating the transcriptomics data file (default NULL)}

\item{phospho_file}{a string (optional) indicating the phosphoproteomics data file (default NULL)}

\item{tf_params}{A list of parameters for TFEA analysis. Available parameters are:
\itemize{
\item \code{omic_data}: dataframe of transcriptomic data for the analysis.
\item \code{analysis}: analysis type; default: \code{'tfea'}.
\item \code{organism}: reference species; default: \code{'human'}.
\item \code{reg_minsize}: minimum regulons group size; default: \code{10}.
\item \code{exp_sign}: boolean, indicates whether to consider the sign of the expression; default: \code{FALSE}.
\item \code{collectri}: boolean, indicates whether to collect regulator-target relationships; default: \code{FALSE}.
\item \code{hypergeom_corr}: boolean, whether to correct results for hypergeometry; default: \code{TRUE}.
\item \code{GO_annotation}: boolean, whether to include GO annotations; default: \code{TRUE}.
\item \code{correct_proteomics}: boolean, whether to correct with proteomic data; default: \code{TRUE}.
\item \code{prot_df}: supporting proteomic data frame.
\item \code{custom}: boolean, whether to use custom regulons; default: \code{FALSE}.
\item \code{custom_path}: path to custom regulons file; default: \code{NULL}.
}}

\item{kin_params}{A list of parameters for the KSEA analysis. Available parameters are:
\itemize{
\item \code{omic_data}: dataframe of phosphoproteomic data for analysis.
\item \code{analysis}: analysis type; default: \code{'ksea'}.
\item \code{organism}: reference species; default: \code{'human'}.
\item \code{reg_minsize}: minimum size of the regulator group; default: \code{5}.
\item \code{exp_sign}: boolean, whether to consider the sign of the expression; default: \code{FALSE}.
\item \code{integrated_regulons}: boolean, whether to consider integrated regulators; default: \code{TRUE}.
\item \code{hypergeom_corr}: boolean, whether to correct the results for hypergeometry; default: \code{TRUE}.
\item \code{GO_annotation}: boolean, whether to include GO annotations; default: \code{TRUE}.
\item \code{correct_proteomics}: boolean, whether to correct with proteomic data; default: \code{TRUE}.
\item \code{prot_df}: supporting proteomic data frame.
\item \code{custom}: boolean, whether to use custom annotations; default: \code{FALSE}.
\item \code{custom_path}: path to the custom annotation file; default: \code{NULL}.
}}

\item{phosphoscore_params}{A list of parameters for the calculation of \emph{phosphoscore}. Available parameters are:
\itemize{
\item \code{phosphoproteomic_data}: dataframe of phosphoproteomic data for the calculation of \emph{phosphoscore}.
\item \code{organism}: reference species; default: \code{'human'}.
\item \code{activatory}: boolean, whether to consider activatory changes; default: \code{TRUE}.
\item \code{GO_annotation}: boolean, whether to include GO annotations; default: \code{TRUE}.
\item \code{custom}: boolean, whether to use custom annotations; default: \code{FALSE}.
\item \code{custom_path}: path to the custom annotation file; default: \code{NULL}.
}}

\item{phosphoscore_noseqwin_params}{A list of parameters for calculating \emph{phosphoscore} without \emph{sequence window}.
The available parameters are the same as those in \code{phosphoscore_params} and are overridden if specified.}

\item{output_dir}{a string indicating the output directory that contains results files}
}
\value{
No return value. The final results are saved in Excel files in the output directory.
}
\description{
This wrapper to SignalingProfiler preotein activity inference, extracts and analyzes the protein activity for a specific patient starting from omic data: transcriptomics, proteomics and phosphoproteomics.
It combines footprint based analysis (TFEA and KSEA), with phosphorylation scores (Phosphoscore), generating aggregated results for the patient.
}
\details{
\emph{Results output}:
Final results for the patient will be saved as an xlsx files inside the 'Activities' directory created by the function with this format:
\verb{Activity_patient_\{patient_id\}.xlsx}.
Each file contains these columns:
\itemize{
\item \code{UNIPROT}: protein Uniprot ID.
\item \code{gene_name}: name of the gene.
\item \code{mf}: molecular functions.
\item \code{final_score}: activity score.
\item \code{method}: how final_score is calculated.
}
}
\examples{
tf_params <- list(reg_minsize = 5, collectri = TRUE)
kin_params <- list(exp_sign = TRUE, GO_annotation = FALSE)
extract_protein_activity(prot_file = "Prot_patients/Prot_Patient_CPT000814.xlsx",
                         trans_file = "Transc_patients/Transc_Patient_CPT000814.xlsx",
                         phospho_file = "Phospho_patients/Phospho_Patient_CPT000814.xlsx",
                         tf_params,
                         kin_params,
                         output_dir = "Activities")

}
