% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_signatures.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extract signaling-driven transcriptomics signature from communities}
\usage{
extract_signatures(
  base_path = "./vignette/Communities/output_communities/",
  transcriptomics_file = "./PatientProfiler_processed_input/Transcriptomics_updated.csv",
  output_file = "Anova_result.csv",
  padj_thres = 0.01,
  diff_thres = 0.7,
  mean_exp_clus_thres = 0,
  max_val = 50
)
}
\arguments{
\item{base_path}{Character. Path to the directory containing community folders.}

\item{transcriptomics_file}{Character. Path to the transcriptomic data file.}

\item{output_file}{Character. Name of the output file containing all results.}

\item{padj_thres}{Numeric. Adjusted p-value threshold for filtering significant results. Default: \code{0.01}.}

\item{diff_thres}{Numeric. Minimum difference threshold for gene expression filtering. Default: \code{0.7}.}

\item{mean_exp_clus_thres}{Numeric. Minimum mean expression threshold for gene filtering. Default: \code{0}.}

\item{max_val}{Integer. Maximum number of genes to include in each community-specific signature. Default: \code{50}.}
}
\value{
A data frame containing the results of the transcriptomic signature for each community.
The function also saves an Excel file with the anova results and generates a "Signatures" directory where filtered signature files for each community are stored.
}
\description{
This function analyzes transcriptomic data by performing ANOVA-Tukey on gene expression levels and then generating community-specific gene expression signatures
}
\examples{
extract_signatures <- function(base_path = "./vignette/Communities/output_communities/",          
transcriptomics_file = "./PatientProfiler_processed_input/Transcriptomics_updated.csv", 
output_file = "Anova_result.csv",
padj_thres = 0.01,
diff_thres = 0.5,
mean_exp_clus_thres = 0,
max_val = 50)
}
