% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_df_to_list}
\alias{from_df_to_list}
\title{From df to list}
\usage{
from_df_to_list(df_cols)
}
\arguments{
\item{df_cols}{df with one row and many cols}
}
\value{
list with cols as elements
}
\description{
To use to convert a mutation matrix to a list for vanilla CARNIVAL
}
\examples{

patient_id <- 'CPT000814'
mutations_df <- read_csv('./mutations.csv')
mut_pat <- mutations_df[mutations_df$Patient_ID == patient_id,]
from_df_to_list(mut_pat)

}
