% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_communities.R
\name{generate_communities}
\alias{generate_communities}
\title{Generate communities from patient-specif mechanistic models}
\usage{
generate_communities(dir_path, network_dir, output_dir, local = FALSE)
}
\arguments{
\item{dir_path}{Character. Path to the directory containing the patients stratification.tsv in Subtype}

\item{network_dir}{Character. Path to the directory containing patinet-specific mechanistic models obtained in Step 3}

\item{output_dir}{Character. Path to the directory containing the output of this function}

\item{local}{Boolean.}
}
\value{
A directory for each community containing nodes table, edges table e patients stratification.
}
\description{
Generate communities from patient-specif mechanistic models
}
\examples{
NULL


}
