% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PKN.R
\name{get_PKN}
\alias{get_PKN}
\title{get_PKN}
\usage{
get_PKN(PKN_params = list(), omics_list = NULL)
}
\arguments{
\item{PKN_params}{A list of parameters for PKN selection. Available parameters are:
\itemize{
\item \code{preprocess}: Boolean, whether to remove interactions involving entities not detected in at least one omic data; default \code{TRUE}.
\item \code{direct}: Boolean, whether keep only direct interactions; default \code{TRUE}.
\item \code{organism}: string, available are 'human' or 'mouse'; default \code{human}.
\item \code{with_atlas}: Boolean, whether integratins Ser/Thr and Tyr atlas inferred interactions; default \code{FALSE}.
\item \code{custom}: Boolean, whether the user wants to provide a custom PKN;  default \code{FALSE}.
\item \code{custom_path}: string, path to the custom PKN;  default \code{FALSE}.
}}

\item{omics_list}{A list of omics data to use to preprocess PKN.
Each omic file should contain the \code{gene_name} column reporting the Gene Symbol.}
}
\value{
Returns a dataframe representing the interactions of in the PKN
}
\description{
Selects the desired SIGNOR and PhosphoSitePlus built-in interactome
}
\details{
\emph{Results output}
A dataframe representing SIGNOR and PhosphoSitePlus built-in interactome of SignalingProfiler package
}
\examples{
omics_list <- list(transcriptomics, proteomics)
PKN_params = list(preprocess = TRUE, direct = TRUE, organism = 'human')
}
