% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igraphToSif}
\alias{igraphToSif}
\title{igraphToSif}
\usage{
igraphToSif(inGraph, outfile = "output.sif", edgeLabel = "label")
}
\arguments{
\item{inGraph}{igraph object}

\item{outfile}{name of output file}

\item{edgeLabel}{column containing the SIGN information}
}
\value{
no return.
igraphToSif(naive_network, outfile='naive.sif', edgeLabel='sign')
}
\description{
converts an igraph object to SIF file
}
