% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{initialize_net_default_params}
\alias{initialize_net_default_params}
\title{initialize_net_default_params}
\usage{
initialize_net_default_params(output_dir)
}
\arguments{
\item{output_dir}{user-provided network building directory}
}
\value{
it returns a list of 5 lists for each internal function:
\itemize{
\item \verb{$PKN_options}: parameters list for SignalingProfiler PKN choice of 7 elements:
\itemize{
\item \code{preprocess}: Boolean, whether to remove interactions involving entities not detected in at least one omic data; default \code{TRUE}.
\item \code{direct}: Boolean, whether keep only direct interactions; default \code{TRUE}.
\item \code{organism}: string, available are 'human' or 'mouse'; default \code{human}.
\item \code{with_atlas}: Boolean, whether integratins Ser/Thr and Tyr atlas inferred interactions; default \code{FALSE}.
\item \code{custom}: Boolean, whether the user wants to provide a custom PKN;  default \code{FALSE}.
\item \code{custom_path}: string, path to the custom PKN;  default \code{FALSE}.
}
\item \verb{$naive_options}: parameters list for SignalingProfiler naive network layers of 4 elements:
\itemize{
\item \code{layers}: integer, specifiying the number of layers > 0 and < 4; default  \code{2}
\item \code{max_length} a vector of int or an int, length == layers; default  \code{c(1,4)}
\item \code{connect_all} Boolean, whether connecting intermediate nodes of shortest path; default  \code{TRUE}
\item \code{naive_path} path of the folder where naive network files are saved; default \code{./Network_output/Naive_}
}
\item \verb{$carnival_options}: parameters list for SignalingProfiler CARNIVAL implementations of 4 elements:
\itemize{
\item \code{solver}: string, specifiying the ILP solver for CARNIVAL; default  \code{cplex}
\item \code{carnival_type}: string, specifying the CARNIVAL implementation among 'inverse', 'vanilla_one_shot', 'vanilla_two_shots', 'vanilla_three_shots';
\item \code{opt_path}: path to the folder were optimized networks files are saved; default \code{./Network_output/Opt_}
\item \code{carnival_params}: a list of CARNIVAL parameters, set in \emph{SignalingProfiler::default_CARNIVAL_options(solver)}
}
\item \verb{$phenoscore_options}: parameters list for SignalingProfiler PhenoScore computation of 3 elements:
\itemize{
\item \code{pheno_path}:  path to the folder were optimized networks files are saved; default \code{./Network_output/Pheno_}
\item \code{pheno_distance_table}: output of \emph{proxpath_preprocessing} function; default \code{NULL}.
\item \code{phenoscore_params}: list of 9 elements for PhenoScore computation:
\itemize{
\item \code{path_length}: max length between proteins and phenotypes; default \code{3}.
\item \code{stat}: ProxPath z-score statistics for significantly close phenotypes; default \code{mean}.
\item \code{zscore_threshold}: ProxPath z-score threshold for significantly close phenotypes; default \code{-1.96}.
\item \code{nrandom}: ProxPath number of randomization of input proteins length;  default \code{1000}.
\item \code{pvalue_threshold}: ProxPath t-test p-value adjusted randomization threshold;  default \code{0.05}.
\item \code{remove_cascade}: in linear cascade of proteins regulating a phenotype, consider only the most upstream one in PhenoScore computation; default \code{TRUE}.
\item \code{node_idx}: whether weighting protein activity contribution to phenotypes' regulation based on number of paths; default  \code{FALSE}.
\item \code{use_carnival_activity}: whether consider all proteins in the network inferred by carnival (TRUE) or only the experimentally-derived ones (FALSE); default \code{FALSE}.
\item \code{create_pheno_network}: whether connect model proteins to phenotypes; default \code{TRUE}.
}
}
\item \verb{$format_options}: parameters list for manipulation of proteins-to-phenotypes models of 3 elements:
\item \code{optimize_on_phenotypes}: Boolean, whether keep only proteins-to-phenotypes edges coherent with the inferred phenotypes activity; default, \code{FALSE}.
\item \code{circuits_params}: a list of parameters for functional circuits creation:
\itemize{
\item \code{k}: max path length between proteins an phenotypes in the model; if \code{circuits_params$k=-1} no circuits will be created;
\item \code{start_to_top}: Boolean, whether starting from phenotypes looking for upstream regulators or from sources; default, \code{TRUE}.
}
\item \code{vis_cytoscape}: Boolean, whether using \emph{RCy3} package to visualize networks using SignalingProfiler style (if TRUE, ensure to have Cytoscape open); default, \code{FALSE}.
}
}
\description{
set default PatientProfiler network building parameters
}
