% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{install_pp_py}
\alias{install_pp_py}
\title{install_pp_py}
\usage{
install_pp_py(
  ...,
  envname = "r-patientprofiler",
  new_env = identical(envname, "r-patientprofiler")
)
}
\arguments{
\item{...}{flexible arguments}

\item{envname}{"r-patientprofiler"}

\item{new_env}{used-specified name for the environment}
}
\value{
no return, r-patientprofiler virtual environment created
}
\description{
This function must be executed the first time PatientProfiler is installed
}
\examples{

library(PatientProfiler)
install_pp_py()

}
