% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseplot.R
\name{make_baseplot}
\alias{make_baseplot}
\title{Make the final baseplot}
\usage{
make_baseplot(data, ...)
}
\arguments{
\item{...}{list of additional arguments}

\item{dt}{data frame}
}
\description{
Make the final baseplot
}

