% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_peptide.R
\name{modify_peptide}
\alias{modify_peptide}
\title{modify_peptide}
\usage{
modify_peptide(phospho_df, mult_col, peptide_col, site_col)
}
\arguments{
\item{phospho_df}{dataframe, phosphoproteomics data}

\item{mult_col}{integer, number of the Multiplicity column}

\item{peptide_col}{name of the column containing the peptide sequence}

\item{site_col}{integer, number  of the Site column}
}
\value{
phosphoproteomics dataframe with just the phosphosite's amminoacid in lowercase
}
\description{
modify_peptide
}
\examples{

#' sample_df <- data.frame(
  Gene_name = c("ACIN1", "ZZEF1", "AAK1")
  Peptide = c("sRsRsTPVRR", "LLsFRsMEEAR", "SATTTPST*PR"),
  Multiplicity = c(3, 2, 1),
  Site = c("S245", "S30", "T100"),
  Value = c(0.1, 0.2, 0.3)
)

modified_df <- modify_peptide(sample_df, mult_col = 3, peptide_col = 2, site_col = 4)
}
