% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_preparation.R
\name{omics_preparation}
\alias{omics_preparation}
\title{Omics Data Preparation}
\usage{
omics_preparation(
  df_tr_updated = NULL,
  df_pr_updated = NULL,
  df_ph_updated = NULL,
  transc_dir_name = "Transc_patients",
  prot_dir_name = "Prot_patients",
  phospho_dir_name = "Phospho_patients"
)
}
\arguments{
\item{df_tr_updated}{dataframe containing transcriptomics data. Must include 'gene_ID' and 'gene_name', with patient sample columns.}

\item{df_pr_updated}{dataframe containing proteomics data. Must include 'UNIPROT', 'gene_name', and patient sample columns.}

\item{transc_dir_name}{name of the directory for saving transcriptomics data outputs per patient.}

\item{prot_dir_name}{name of the directory for saving proteomics data outputs per patient.}

\item{phospho_dir_name}{name of the directory for saving phosphoproteomics data outputs per patient.}
}
\description{
This function prepares transcriptomics, proteomics, and phosphoproteomics data for each patient
by creating separate CSV files for individual patient data. It performs various preprocessing steps,
including calculation of significance, log-p-values, and reordering of columns. The output files
are saved in separate directories for each type of omics data.
}
\examples{
omics_preparation(df_tr_updated = transcriptomics_data,
                  df_pr_updated = proteomics_data,
                  df_ph_updated = phosphoproteomics_data,
                  transc_dir_name = "Transc_patients",
                  prot_dir_name = "Prot_patients",
                  phospho_dir_name = "Phospho_patients")

}
