% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_update.R
\name{omics_update}
\alias{omics_update}
\title{omics_update}
\usage{
omics_update(
  df_tr = NULL,
  df_pr = NULL,
  df_ph = NULL,
  threshold = 80,
  sw_len = 7,
  uniprot_idx = NULL,
  pep_col_name = NULL,
  imp_method = "pmm",
  zscore = "TRUE",
  zmethod = "column",
  metric = "median",
  output_dir = "PatientProfiler_processed_input"
)
}
\arguments{
\item{df_tr}{dataframe containing transcriptomics data.}

\item{df_pr}{dataframe containing proteomics data.}

\item{df_ph}{dataframe containing phosphoproteomics data.}

\item{threshold}{numeric, it specifies the threshold percentage of zeros in each row the transcriptomic dataframe. If there are >= threshold percentage of
zeros in a row, the row will be removed (default: 80).}

\item{sw_len}{integer, the sequence window length. It could be 7 or 15 (default: 7).}

\item{uniprot_idx}{optional integer indicating the column index for UNIPROT IDs to be added at the uniprot column retrieved with AnnotationDbi (default is NULL).}

\item{pep_col_name}{string, it indicates the name of the peptide sequence column (if it's present) in your phosphoproteomic dataframe (default = "Peptide")}

\item{imp_method}{string, the method to use for imputation (default: "pmm", but you can choose between "pmm", "norm", "norm.nob",
"regression", "ri", "logreg", "polyreg", "predictive", "polr", "sample", "cart", "knn", "rf").}

\item{zscore}{logical, whether to perform Z-score normalization (default: "TRUE").}

\item{zmethod}{string, specifies whether Z-score normalization is performed by "row" or "column" (default: "column").}

\item{metric}{string, the centering metric for Z-score normalization. Options are "median" (default) or "mean".}

\item{output_dir}{a string indicating the updated output folder.}
}
\value{
This function updates the following global variables:

transctriptomics_updated -> A data frame containing the updated transcriptomics data.
proteomics_updated -> A data frame containing the updated proteomics data.
phosphoproteomics_updated -> A data frame containing the updated phosphoproteomics data.
}
\description{
This function updates transcriptomics, proteomics, and phosphoproteomics datasets by performing
preprocessing steps such as filtering, imputation, and Z-score normalization.
Dataframes must contain the following required columns:
\itemize{
\item df_tr: 'gene_name' and sample columns.
\item df_pr: 'gene_name' and sample columns.
\item df_ph: 'gene_name', 'Site' (containing phosphorylation sites in the format: T161), and 'Peptide'
(containing the peptide sequence; the name and the presence of the peptide column is optional).
}
}
\examples{
transcriptomics_data <- data.frame(matrix(rnorm(100), nrow = 10))
proteomics_data <- data.frame(matrix(rnorm(100), nrow = 10))
phosphoproteomics_data <- data.frame(matrix(rnorm(100), nrow = 10))

omics_update(
  df_tr = transcriptomics_data,
  df_pr = proteomics_data,
  df_ph = phosphoproteomics_data,
  sw_len = 7,
  uniprot_idx = NULL,
  pep_col_name = NULL,
  imp_method = "pmm",
  zscore = TRUE,
  zmethod = "column",
  metric = "median"
)

}
