% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_network_with_carnival.R
\name{optimize_network_with_carnival}
\alias{optimize_network_with_carnival}
\title{optimize_network_with_carnival}
\usage{
optimize_network_with_carnival(
  sources,
  activities,
  naive_network,
  phosphoproteomics = NULL,
  save_all_files = FALSE,
  network_params = list()
)
}
\arguments{
\item{sources}{dataframe of perturbed nodes (e.g., patient mutations), starting point of the model}

\item{activities}{dataframe of proteins inferred activity to connect to the starting point}

\item{naive_network}{igraph object representing the naive network}

\item{phosphoproteomics}{dataframe of phosphoproteomics; default: \code{NULL} (no phosphoproteomics mapping on network).}

\item{save_all_files}{Boolean, if TRUE it will save 13 files per patient, FALSE, just 4; default \code{FALSE}.}

\item{network_params}{check \emph{initialize_net_default_params} documentation, in the \verb{$carnival_options} part.}
}
\value{
A list of three elements:
\itemize{
\item \code{igraph_network}: optimized network in igraph format
\item \code{nodes_df}: dataframe of the nodes in the optimized network with the following attributes:
\itemize{
\item \code{gene_name}: HUGO gene symbol of each node
\item \code{UNIPROT}: UNIPROT ID of each node
\item \code{carnival_activity}: node activity predicted by CARNIVAL in the optimization process (usually equal to final_score)
\item \code{final_score}: node activity predicted by extract_protein_activity from experimental data
\item \code{method}: algorithm of the activity; can be CARNIVAL, VIPER (footprint-based analysis), or PhosphoScore
\item \code{discordant}: Boolean, whether \code{carnival_activity} and \code{final_score} have opposite signs
\item \code{mf}: molecular function of each node (rec, kin, phos, other, tf)
}
\item \code{edges_df}: dataframe of the edges in the optimized network with the following attributes:
\itemize{
\item \code{soource}: HUGO gene symbol of starting node
\item \code{target}: HUGO gene symbol of target node
\item \code{sign}: sign of the interactions (1, -1)
\item \code{carnival_weight}: interaction confidence in CARNIVAL optimization (0-100)
\item \code{direct}: Boolean, whether the interaction is direct
\item \code{mechanism}: mechanism associated to the interaction in SIGNOR
\item \code{aminoacid}: if \code{phosphoproteomics} is provided, represents putative phosphosites modulated in experimental data  modified by the interaction
\item \code{is_quantified}: if \code{phosphoproteomics} is provided, when TRUE the mapped phosphosite modulation is coherent with interaction sign, but not necessarily significant
\item \code{is_significant}:  if \code{phosphoproteomics} is provided, when TRUE the mapped phosphosite modulation is coherent with interaction sign and significant
\item \code{FC}: if \code{phosphoproteomics} is provided, reports the phosphosite's modulation in experimental data
}
}
}
\description{
Wrapper function to the four implementations of vanilla and inverse CARNIVAL in SignalingProfiler
to retrieve sign-coherent interactions with inferred protein activities
}
\details{
\emph{Results output}
An optimized network with sign-coherent edges as:
\itemize{
\item a \emph{list} of igraph_network, nodes and edges table;
\item creates internally an \emph{RDS file} and a \emph{SIF file} of the optimized network at \code{output_dir} directory
}
}
\examples{
network_params <- initialize_net_default_params('./Network_output/')
network_params$carnival_options <- list(solver = 'cplex', carnival_type = 'inverse')

omics_list <- list(transcriptomics, proteomics, phosphoproteomics)
PKN = get_PKN(omics_list = omics_list)
naive_network = create_naive_network(PKN, sources, activities, naive_network_parameters = list(layers = 2, max_length = c(1,4)))
carnival_network = optimize_network_with_carnival(sources, activities, phosphoproteomics, network_params = network_params)
}
