% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patientplot.R
\name{plot_patient}
\alias{plot_patient}
\title{Add a patient's data to a baseplot}
\usage{
plot_patient(baseplot, patient_data, title = "", aes_map = aes_map(),
  vline_map = vline_aes(), segment_size = 2.5)
}
\arguments{
\item{baseplot}{the plot to add the geoms}

\item{patient_data}{an individual patient's data}

\item{title}{title of the plot}

\item{aes_map}{a list of the aesthetic mappings for the plot}

\item{segment_size}{width of the segments}

\item{vline}{a list of the aesthetics for vline}
}
\description{
Add a patient's data to a baseplot
}

