% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_link_phenotypes.R
\name{proxpath_preprocessing}
\alias{proxpath_preprocessing}
\title{proxpath_preprocessing}
\usage{
proxpath_preprocessing(proteomics, phosphoproteomics)
}
\arguments{
\item{proteomics}{dataframe of proteomics; default: \code{NULL}.}

\item{phosphoproteomics}{dataframe of phosphoproteomics; default: \code{NULL}.}
}
\value{
\code{pheno_distances_table} a table reporting the distance between each protein-phenotype pair in SIGNOR
}
\description{
This function calls a Python script in SignalingProfiler that downloads proteins-to-phenotypes interactions in SIGNOR
and preprocess it according to (phospho)proteomics data, removing interactions not quantified experimentally
If no omics are provided, the function return the built-in object in SignalingProfiler
}
\examples{
pheno_distances_table <- proxpath_preprocessing(proteomics = NULL, phosphoproteomics = phosphoproteomics)

}
