% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_phosphoproteomics.R
\name{remove_duplicates_phosphoproteomics}
\alias{remove_duplicates_phosphoproteomics}
\title{remove_duplicates_phosphoproteomics}
\usage{
remove_duplicates_phosphoproteomics(df)
}
\arguments{
\item{df}{dataframe (phosphoproteomics data). It must include 'UNIPROT', 'gene_name', 'aminoacid}
}
\value{
a cleaned data frame with aggregated metadata and averaged values.
}
\description{
This function processes a phosphoproteomics data frame by removing duplicates,
aggregating metadata, and calculating average values for patient columns.
}
\examples{
phosphoproteomics_data <- data.frame(
  gene_name = c("Gene1", "Gene1", "Gene2"),
  aminoacid = c("A", "A", "B"),
  position = c(1, 1, 2),
  sequence_window = c("AAACCC", "AAACCC", "BBBCCC"),
  UNIPROT = c("P12345", "P12345", "P67890"),
  patient_1 = c(0.5, 0.6, 0.7),
  patient_2 = c(0.4, 0.3, 0.8)
)

cleaned_data <- remove_duplicates_phosphoproteomics(phosphoproteomics_data)
head(cleaned_data)

}
