% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_proteomics.R
\name{remove_duplicates_proteomics}
\alias{remove_duplicates_proteomics}
\title{Remove Duplicate Entries in Proteomics Data}
\usage{
remove_duplicates_proteomics(df)
}
\arguments{
\item{df}{dataframe (proteomics data). It must include at least the 'gene_name' and 'UNIPROT' columns.}
}
\value{
a cleaned dataframe where each \code{gene_name} has a unique entry, and
associated \code{UNIPROT} identifiers are aggregated.
}
\description{
This function processes a proteomics data frame to remove duplicate entries
based on the \code{gene_name} column. It ensures that for each gene, a unique entry
is retained, and all associated \code{UNIPROT} identifiers are combined into a single string,
separated by semicolons.
}
\examples{
cleaned_proteomics <- remove_duplicates_proteomics(proteomics_data)
head(cleaned_proteomics)

}
