% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_invalid_site.R
\name{remove_invalid_site}
\alias{remove_invalid_site}
\title{remove_invalid_site}
\usage{
remove_invalid_site(phospho_df, site_col)
}
\arguments{
\item{phospho_df}{tibble, the phosphoproteomics dataframe}

\item{site_col}{integer, the column number of the Site}
}
\value{
phosphoproteomics dataframe with no uncertain site
}
\description{
remove_invalid_site
}
\examples{

sample_df <- data.frame(
  Gene_name = c("UNKL", "ATF7", "SAMD1"),
  Site = c("S100T345", "T56", "Y389"),
  Value = c(0.1, 0.2, 0.3)
)

cleaned_df <- remove_invalid_site(sample_df, site_col = 2)
}
