% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiplicity.R
\name{remove_multiplicity_by_prefix}
\alias{remove_multiplicity_by_prefix}
\title{remove_multiplicity}
\usage{
remove_multiplicity_by_prefix(phospho_df, peptide_col, mult_col, gn_idx)
}
\arguments{
\item{phospho_df}{tibble, the phosphoproteomics dataframe}

\item{peptide_col}{name of the column containing the peptide sequence}

\item{mult_col}{integer, the number of the Multiplicity column}

\item{gn_idx}{integer, number of the Gene name column}
}
\value{
phosphoproteomics dataframe with duplicates removed based on multiplicity
}
\description{
remove_multiplicity
}
\examples{

#' sample_df <- data.frame(
  Gene_name = c("ACIN1", "ZZEF1", "AAK1")
  Peptide = c("sRsRsTPVRR", "LLsFRsMEEAR", "SATTTPSG*PR"),
  Multiplicity = c(3, 2, 1),
  Value = c(0.1, 0.2, 0.3)
)

cleaned_df <- remove_multiplicity(sample_df, peptide_col = "Peptide", mult_col = 3, gn_idx = 1)
}
