% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_coding.R
\name{retrieve_coding}
\alias{retrieve_coding}
\title{Filter for coding genes}
\usage{
retrieve_coding(df)
}
\arguments{
\item{df}{tibble, a transcriptomics dataframe containing gene names.}

\item{gn_col}{string, the name of the column in \code{df} containing gene symbols to match against coding genes (default: "gene_name").}
}
\value{
a filtered dataframe containing only rows with coding genes.
}
\description{
This function filters a data frame to retain only rows corresponding to coding genes based on the \code{org.Hs.eg.db} database.
}
\examples{
# Example usage:
filtered_df <- retrieve_coding(my_df, gn_col = "GeneSymbol")
head(filtered_df)

}
