% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_primary_gene_name_and_sequence.R
\name{retrieve_primary_gene_name}
\alias{retrieve_primary_gene_name}
\title{Retrieve Primary Gene Name from UNIPROT Database}
\usage{
retrieve_primary_gene_name(
  omic_df = br_pr,
  gn_idx = 1,
  uniprot_idx = NULL,
  organism = "human",
  sequence = TRUE
)
}
\arguments{
\item{omic_df}{dataframe containing omics data. It must have a column with gene symbols.}

\item{gn_idx}{integer, the column index for gene names in the omics data frame.}

\item{uniprot_idx}{optional integer indicating the column index for UNIPROT IDs to be removed (default is NULL).}

\item{organism}{string, it indicates the organism. Options are 'human' (default) or 'mouse'.}

\item{sequence}{logical, indicating whether to retain the protein sequence information in the result (default is TRUE).}
}
\value{
A data frame with updated gene names and corresponding UNIPROT IDs. If \code{sequence = TRUE}, the sequence information is also included.
}
\description{
This function updates a given omics data frame by retrieving the primary gene name
and associated UNIPROT ID for each gene symbol. It queries the UNIPROT database
for human or mouse gene symbols and returns the most relevant data based on sequence
modifications and review status.
}
\examples{
# Example usage:
updated_omic_df <- retrieve_primary_gene_name(omic_df, gn_idx = 2, organism = 'human')
head(updated_omic_df)
}
