% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseplot.R
\name{scale_baseplot}
\alias{scale_baseplot}
\title{Map aesthetic scales for a patient profile}
\usage{
scale_baseplot(p, color.values = brewer_pal("qual", pal =
  2)(length(levels(p$data$color))), color.guide = FALSE,
  size.values = (1:length(levels(p$data$size)))/length(levels(p$data$size)),
  size.guide = FALSE, shape.values = 1:length(levels(p$data$shape)),
  shape.guide = FALSE, fill.values = brewer_pal("qual", pal =
  2)(length(levels(p$data$color))), fill.guide = FALSE)
}
\arguments{
\item{p}{a ggplot object}
}
\description{
Map aesthetic scales for a patient profile
}

