% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_signature_on_dataset.R
\name{test_signature_on_dataset}
\alias{test_signature_on_dataset}
\title{test_signature_on_dataset}
\usage{
test_signature_on_dataset(
  gene_signatures,
  transcriptomic_data,
  clinical_data,
  NES_filter = TRUE,
  pval_cutoff = 0.05,
  palette = "npg",
  time_threshold = NULL,
  exclude_communities = NULL,
  enrichment_result = NULL
)
}
\arguments{
\item{gene_signatures}{list of named vector specifying the genes for each signature}

\item{transcriptomic_data}{dataframe, where Gene Symbols are specified in \emph{gene_name} column}

\item{clinical_data}{dataframe, with \emph{patient_id}, \emph{serial_time} (days_to_death or days_to_last_followup), \emph{status} (0 alive, 1 decesead)}

\item{NES_filter}{Boolean, if \emph{TRUE} only positive enrichment is considered}

\item{pval_cutoff}{padj cutoff for fgsea; default \code{0.05}.}

\item{time_threshold}{integer, \emph{serial_time} days to filter \emph{clinical_data} based on; default \code{NULL}.}

\item{exclude_communities}{character vector, a vector matching the names of gene_signatures, to exclude specific groups from KM; default \code{NULL}.}

\item{enrichment_result}{fgsea result; default \code{NULL}.}
}
\value{
A list of five elements:
\itemize{
\item \emph{enrichment_results} fgsea enrichment result
\item \emph{hazard_ratio} hazard ration from survival analysis
\item \emph{p_value} chi-squared test pvalue
\item \emph{adjusted_p_value} chi-squared test BH adjusted pvalue
\item \emph{plot} Kaplan-Meier plot
}
}
\description{
Run fgsea on user-specified set of \emph{gene_signatures} using transcriptomics data.
If \emph{enrichment_result} is provided this step is skipped.
If \emph{clinical_data} are provided, it draws a Kaplain-Meier curve assigning each patient
to the most-enriched signature (lowest padj and higher NES).
}
\examples{
#Read signatures from Signatures folder 
test_signature_on_dataset(gene_signatures = gene_signatures,
                         transcriptomic_data = transc,
                         clinical_data = clinical,
                         NES_filter = TRUE,
                         pval_cutoff = 0.05,
                         palette = 'npg',
                         time_threshold = NULL,
                         exclude_communities = NULL,
                         enrichment_result = NULL)
}
