% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseplot.R
\name{theme_baseplot}
\alias{theme_baseplot}
\title{Theme a patient profile}
\usage{
theme_baseplot(p)
}
\arguments{
\item{p}{a ggplot object}
}
\description{
Adds a theme to a ggplot object. Not very useful at the moment
}

