% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_update.R
\name{transcriptomics_update}
\alias{transcriptomics_update}
\title{transcriptomics_update}
\usage{
transcriptomics_update(
  df_tra,
  threshold = 80,
  zscore = TRUE,
  zmethod = "column",
  metric = "median"
)
}
\arguments{
\item{df_tra}{tibble, a data frame containing the transcriptomics data.}

\item{threshold}{numeric, it specifies the threshold percentage of zeros in each row the transcriptomic dataframe. If there are >= threshold percentage of
zeros in a row, the row will be removed.}

\item{zscore}{boolean, indicates whether to calculate z-scores for the data.
If TRUE, z-scores are calculated; if FALSE, a data frame will be returned without z-score calculation.}

\item{zmethod}{string, it specifies the method for calculating z-scores. It can either be "row" to compute z-scores for each row,
or "column" to compute z-scores for each column of the data.}

\item{metric}{string, it indicates the method for centering the data when z-scores are computed.
It could be "mean" or "median".}
}
\value{
tibble, it contains the filtered and processed transcriptomics data.
}
\description{
This function processes and filters a transcriptomics dataframe based on user-defined parameters.
It can filter rows based on the percentage of missing values (NA), and optionally compute z-scores
for the data either by rows or columns.
}
\examples{

df_processed <- transcriptomics_update(df_tra, zscore = TRUE, zmethod = "median", metric = "mean")

df_cleaned <- transcriptomics_update(df_tra, zscore = FALSE, zmethod = "median", metric = "mean")
}
