% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_position_aa.R
\name{update_position_aa}
\alias{update_position_aa}
\title{update_position_aa}
\usage{
update_position_aa(phospho_df, sw_col, sequence_col, peptide_col)
}
\arguments{
\item{phospho_df}{tibble, the phosphoproteomics dataframe}

\item{sw_col}{integer, the number of the sequence window column}

\item{sequence_col}{name of the column containing the protein sequence}

\item{peptide_col}{name of the column containing the peptide sequence}
}
\value{
phosphoproteomics dataframe with the additional columns 'position' and 'aminoacid' of each phosphosite
}
\description{
update_position_aa
}
\examples{

sample_df <- data.frame(
  Gene_name = c("ACIN1", "ZZEF1", "AAK1"),
  sequence_window = c("HSRSRSRSTPVRD", "LRLLSFRSMEEARLV", "ATTTPSGSPRTSQQN"),
  Sequence = c("MDLEGDRNGGAKSRSRSTPVRKKNFFKLN", "MLENLQKHSTLLSFRSMEEARIGEEMSQNSFIKQY", "YKPQKGLSATTTPSTPRNLGMSEQL"),,
  Peptide = c("SRsRsTPVR", "LLsFRsMEEAR", "SATTTPST*PR"),
  Value = c(0.1, 0.2, 0.3)
)

updated_df <- update_position_aa(sample_df, sw_col = 2, sequence_col = 3, peptide_col = 4)
}
