% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_naive_network.R
\name{validate_naive_network_parameters}
\alias{validate_naive_network_parameters}
\title{validate_naive_network_parameters}
\usage{
validate_naive_network_parameters(
  PKN,
  sources,
  activities,
  output_dir,
  naive_network_parameters
)
}
\arguments{
\item{PKN}{dataframe obtained from the function get_PKN representing PKN interactions}

\item{sources}{dataframe of perturbed nodes (e.g., patient mutations), starting point of the model}

\item{activities}{dataframe of proteins inferred activity to connect to the starting point}

\item{output_dir}{string, output naive network directory}

\item{naive_network_parameters}{A list of parameters for naive network construction. Available parameters are:
\itemize{
\item \code{layers}: integer, specifiying the number of layers > 0 and < 4; default  \code{2}
\item \code{max_length} a vector of int or an int, length == layers; default  \code{c(1,4)}
\item \code{connect_all} Boolean, whether connecting intermediate nodes of shortest path; default  \code{TRUE}
\item \code{naive_path} path of the folder where naive network files are saved; default \code{./Network_output/Naive_}
}}
}
\value{
sources_gn
}
\description{
Validation function for naive network parameters to plain error handling
}
