% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{vline_aes}
\alias{vline_aes}
\title{Create an aesthetic mapping for a Vertical Lines}
\usage{
vline_aes(vline1.color = "black", vline1.size = 1,
  vline1.linetype = "solid", vline2.color = "darkgrey", vline2.size = 1,
  vline2.linetype = "dashed", vline3.color = "lightgrey", vline3.size = 1,
  vline3.linetype = "dotted", vline4.color = "blue", vline4.size = 1,
  vline4.linetype = "dotdash", vline5.color = "red", vline5.size = 1,
  vline5.linetype = "twodash")
}
\arguments{
\item{vline1.color}{defaults to 'black',}

\item{vline1.size}{defaults to 1,}

\item{vline1.linetype}{defaults to 'solid',}

\item{vline2.color}{defaults to 'darkgrey',}

\item{vline2.size}{defaults to 1,}

\item{vline2.linetype}{defaults to 'dashed',}

\item{vline3.color}{defaults to 'lightgrey',}

\item{vline3.size}{defaults to 1,}

\item{vline3.linetype}{defaults to 'dotted',}

\item{vline4.color}{defaults to 'blue',}

\item{vline4.size}{defaults to 1,}

\item{vline4.linetype}{defaults to 'dotdash',}

\item{vline5.color}{defaults to 'red',}

\item{vline5.size}{defaults to 1,}

\item{vline5.linetype}{defaults to 'twodash'}
}
\description{
Create an aesthetic mapping for a Vertical Lines
}

