% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_CPTAC_harmonized_data.R
\name{access_harmonized_CPTAC_data}
\alias{access_harmonized_CPTAC_data}
\title{access_harmonized_CPTAC_data}
\usage{
access_harmonized_CPTAC_data(tumors, omics)
}
\arguments{
\item{tumors}{vector of strings with tumor names. The user can choose from these names: "Brca", "Ccrcc", "Coad", "Gbm", "Hnscc", "Lscc",
"Luad", "Ov", "Pdac", "Ucec".}

\item{omics}{vector of strings containing the omics types to extract. The options could be "phospho", "prot" and "transc"
(respectively phosphoproteomics, proteomics and transcriptomics)}
}
\description{
This function provides access to harmonized CPTAC data for 10 tumor types and 3 omics per tumor type. The data has been harmonized using the
PatientProfiler function \emph{omics_update}. Users can access the data to use them in the other functions for protein activity inference and
for network creation. It creates local variables with the dataframe(s) selected by the user.
}
\examples{
# Extract Phosphoproteomic and Proteomic dataframes for Brca and Ccrcc
access_harmonized_CPTAC_data(tumors = c("Brca", "Ccrcc"), omics = c("phospho", "prot"))

# Four local variables, Brca_phospho, Brca_prot, Ccrcc_phospho and Ccrcc_prot will be generated.

}
