% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sequence_window.R
\name{add_sequence_window}
\alias{add_sequence_window}
\title{add_sequence_window}
\usage{
add_sequence_window(phospho_df = df7, peptide_col, seq_len_i = 7, sequence_col)
}
\arguments{
\item{phospho_df}{tibble, the phosphoproteomics dataframe}

\item{peptide_col}{name of the column containing the peptide sequence}

\item{seq_len_i}{integer, the length of the mers, 7 for 15mers}

\item{sequence_col}{name of the column containing the protein sequence}
}
\value{
phosphoproteomics dataframe with the phosphosite around the central position of the Peptide
}
\description{
add_sequence_window
}
\examples{
sample_df <- data.frame(
  Gene_name = c("ACIN1", "ZZEF1", "AAK1"),
  Peptide = c("SRsRsTPVR", "LLsFRsMEEAR", "SATTTPST*PR"),
  Sequence = c("MDLEGDRNGGAKSRSRSTPVRKKNFFKLN", "MLENLQKHSTLLSFRSMEEARIGEEMSQNSFIKQY", "YKPQKGLSATTTPSTPRNLGMSEQL"),,
  Value = c(0.1, 0.2, 0.3)
)

result_df <- add_sequence_window(sample_df, peptide_col = "Peptide", seq_len_i = 7, sequence_col = "Sequence")
}
