% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{aes_map}
\alias{aes_map}
\title{Create an aesthetic mapping for a Patient Profile}
\usage{
aes_map(fileBy = "id", id = "id", y = "y", x = "x", xend = "xend",
  color = "color", shape = color, size = shape, fill = shape,
  vline1 = NULL, vline2 = NULL, vline3 = NULL, vline4 = NULL,
  vline5 = NULL)
}
\arguments{
\item{fileBy}{variable to group files by}

\item{id}{the subject id variable}
}
\description{
Create an aesthetic mapping for a Patient Profile
}

