% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{aes_map}
\alias{aes_map}
\title{Create an aesthetic mapping for a Patient Profile}
\usage{
aes_map(fileBy = "id", id = "id", y = "y", x = "x", xend = "xend",
  group = "group", color = "color", color.values = NULL,
  color.guide = FALSE, shape = "shape", shape.values = NULL,
  shape.guide = FALSE, size = "size", size.values = NULL,
  size.guide = FALSE, fill = "fill", fill.values = NULL,
  fill.guide = FALSE, alpha = "alpha", alpha.values = NULL,
  linetype = "linetype", linetype.values = NULL, label = "label")
}
\arguments{
\item{fileBy}{variable to group files by}

\item{id}{the subject id variable}
}
\description{
Create an aesthetic mapping for a Patient Profile
}

