% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_communities.R
\name{generate_communities}
\alias{generate_communities}
\title{Generate communities from patient-specif mechanistic models}
\usage{
generate_communities(
  dir_path,
  network_dir,
  output_dir,
  t_lower = 4,
  t_upper = 30
)
}
\arguments{
\item{dir_path}{Character. Path to the directory containing the patients_stratification.tsv in Subtype}

\item{network_dir}{Character. Path to the directory containing patient-specific mechanistic models obtained in Step 3}

\item{output_dir}{Character. Path to the directory containing the output of this function}

\item{t_lower}{Numeric. The lower threshold for filtering edges based on their frequency across patients. Default is 4.}

\item{t_upper}{Numeric. The upper threshold for filtering edges based on their frequency across patients.  If not provided, no upper limit is applied, and all edges above \code{t_lower} will be included.}
}
\value{
A directory for each community containing nodes table, edges table e patients stratification.
}
\description{
Generate communities from patient-specif mechanistic models
}
\examples{
generate_communities(dir_path = "input_communities",
network_dir = "Networks_output",
output_dir = "output_communities")
}
