% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_proteomics.R
\name{impute_proteomics}
\alias{impute_proteomics}
\title{impute_proteomics}
\usage{
impute_proteomics(df, start_column = 1, imputation_method = NULL, m = 5)
}
\arguments{
\item{df}{dataframe with missing data.}

\item{start_column}{integer, the first numeric column to consider for imputation (default: 1).}

\item{imputation_method}{string, the imputation method for \code{mice} (default: "pmm").}

\item{m, }{integer, number of multiple imputations. The default is m = 5.}
}
\value{
The imputed data frame.
}
\description{
This function imputes missing values in a numeric data frame using the \code{mice} package and saves the imputed data
to a global variable for user access.
}
\examples{
result <- impute_data(df, start_column = 1, imputation_method = "pmm")
View(imputed_df) # Access the saved variable

}
