% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_update.R
\name{proteomics_update}
\alias{proteomics_update}
\title{proteomics_update}
\usage{
proteomics_update(
  df_pro,
  impute_method = NULL,
  zscore = TRUE,
  zmethod = "column",
  metric = "median"
)
}
\arguments{
\item{df_pro}{dataframe, containing a column with protein names.}

\item{zscore}{boolean, indicates whether to calculate z-scores for the data.
If TRUE, z-scores are calculated; if FALSE, a data frame will be returned without z-score calculation.}

\item{zmethod}{string, it specifies the method for calculating z-scores. It can either be "row" to compute z-scores for each row,
or "column" to compute z-scores for each column of the data.}

\item{metric}{string, it indicates the method for centering the data when z-scores are computed.
It could be "mean" or "median".}

\item{sequence}{A boolean (TRUE or FALSE) indicating whether the dataset includes a column named "sequence".
If TRUE, the function assumes the "sequence" column is present and may process it accordingly.
If FALSE, the function proceeds without expecting the "sequence" column.}
}
\value{
tibble, it contains the filtered and processed proteomics data.
}
\description{
proteomics_update
}
\examples{
prot_updated <- proteomics_update(df, impute_method = "pmm",zscore = TRUE,zmethod = "column",metric = "median")

}
