% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_community.R
\name{visualize_community}
\alias{visualize_community}
\title{Visualize Community Network in Cytoscape}
\usage{
visualize_community(
  community_name,
  comm_folder = "./Communities/output_communities/"
)
}
\arguments{
\item{community_name}{A string specifying the community name (e.g., "community_7").}

\item{comm_folder}{(optional) Path to the folder with resulting community from previous step. Default: \code{"./Communities/output_communities/"}.}
}
\value{
None. The function creates a network visualization in Cytoscape.
}
\description{
This function loads node and edge data for a specified community,
processes the data, creates a graph, and visualizes it in Cytoscape using
the PatientProfiler predefined style.
}
\examples{

visualize_community("community_7")
}
